Tune OneDrive resumable API call to assure the speed and success rate of file upload

+++ scripts/api.sh
@@ -127,18 +127,20 @@ upload_large_file_by_chunks() {
 	send_api_request "create_upload_session" $1
 	color_print "GREEN" "upload_large_file_by_chunks $1, $(get_human_readble_size ${file_size})"
 
-	local retry_count=0; local retry_max=5
+	local retry_count=0; local retry_max=10
 	if [ -z "${error}" ] || [ "${error}" = "null" ]; then 
 		local upload_url=$(echo ${resp} | jq --raw-output '.uploadUrl')	
 
 		# Note: from the docs, size of each chunk MUST be a multiple of 320 KiB (327,680 bytes)
 		# Since the max size of a mp4 file is less than 11MB, setting chunk size as 5.625M
 		# will let a large file be transfered in only 2 chunks	
-		# 320KB*18 = 327680*18 = 5898240 = 5.625M
-		local chunk_index=0; chunk_size=$((327680*10))  # moderate cpu spike
+		# 320KB*18 = 327680*18 = 5898240 = 5.625M (most video files can be done by 2 chunks, however, CPU spike is higher)
+		# 320KB*12 = 327680*12 = 3932160 = 3.932M (most video files can be done by 3 chunks)
+		local chunk_index=0; chunk_size=$((327680*12))  # moderate cpu spike
 		local range_start=0; range_end=0; range_length=0
 
 		local status_code=200 # used to check chunk send error and retry(re-transmission)
+		local whole_file_retry=false
 		while [ $((${chunk_index}*${chunk_size})) -lt $2 ] && [ ${retry_count} -lt ${retry_max} ]; do
 			range_start=$((${chunk_index}*${chunk_size}))
 			range_end=$((${range_start}+${chunk_size}-1))
@@ -151,38 +153,84 @@ upload_large_file_by_chunks() {
 			# do not use dd to copy file data as curl's input, output it to a tmp file instead
 			# dd if="$1" count=1 skip=${chunk_index} bs=${chunk_size} 2> /dev/null | \
 			# -H "Transfer-Encoding: chunked" \
-			echo "upload chunk: ${chunk_index}, length: ${range_length}, bytes: ${range_start}-${range_end} "
+			echo "$(date +"%F %H:%M:%S") upload chunk: ${chunk_index}, length: ${range_length}, bytes: ${range_start}-${range_end}"
 
 			# by writing to tmp file, the chance of success increases
-			dd if="$1" of="./data/chunk_data" count=1 skip=${chunk_index} bs=${chunk_size} 2> /dev/null 
+			if [ ${retry_count} -eq 0 ]; then 
+				dd if="$1" of="./data/chunk_data" count=1 skip=${chunk_index} bs=${chunk_size} 2> /dev/null 
+			fi 
 			status_code=$(
-				curl -s -k -L -X PUT ${upload_url} \
+				curl -s -k -L -X PUT "${upload_url}" \
 				--data-binary @"./data/chunk_data" \
 				--write-out %{http_code} \
+				--max-time 15 \
 				-H "Content-Type: application/octet-stream" \
+				-H "Accept: application/json; odata.metadata=none" \
 				-H "Content-Length: ${range_length}" \
 				-H "Content-Range: bytes ${range_start}-${range_end}/${2}" \
-				-o /dev/null
-				> /dev/null 2>&1
+				-o /dev/null 
 			)
 
-			if [ ${status_code} -ge 200 ] && [ ${status_code} -lt 205 ]; then
+			# > /dev/null 2>&1
+			if [ \( "${status_code}" == "416" \) -o \( ${status_code} -ge 200 -a ${status_code} -lt 205 \) ]; then 
+				rm ./data/chunk_data  # chunk fragment upload successfully
 				chunk_index=$((${chunk_index}+1))
-				retry_count=0
+				retry_count=0				
 			else 
 				retry_count=$((${retry_count}+1))
+				echo "A re-transmission is required due to status_code: ${status_code}"
 			fi
 			
 			sleep 1
 		done 
 
-		curl -s -k -L -X DELETE ${upload_url}  # delete session after upload
+		curl -s -k -L -X DELETE "${upload_url}"  # delete session after upload
 
-		if [ ${retry_count} -lt ${retry_max} ]; then
+		if [ ${status_code} -eq 201 ]; then
 			color_print "GREEN" "Success: upload_large_file_by_chunks $1"
 		else 
 			color_print "BROWN" "Failed: upload_large_file_by_chunks $1"
 			write_log "Upload file $1 failed with ${retry_count} retry."
 		fi
 	fi
+}  
+
+# Try to get a walk around of OneDrive resumable API which is not very reliable when sending file larger than 4MB: too many timeout, and the speed is very slow.
+# One way is to split the mp4 file into fragments less than 4MB and send them separately by OneDrive small file upload API. 
+# However, the separate fragment file cannot be played without a right tool (like ffmpeg) to split or merge.
+# Spliting video file takes time, and on OneDrive there will be more small files as well. 
+# Thus, this way is not very ideal.
+upload_large_file_by_fragments() {
+	local file_name=$(parse_file_name $1)
+	local file_parent=$(parse_file_parent $1)
+
+	local fragment_path=""; fragment_name=""
+	local chunk_index=0; chunk_size=$((4*1024*1024))  # max is 4MB
+	local range_start=0; range_end=0; range_length=0
+
+	while [ $((${chunk_index}*${chunk_size})) -lt $2 ]; do
+		range_start=$((${chunk_index}*${chunk_size}))
+		range_end=$((${range_start}+${chunk_size}-1))
+		if [ ${range_end} -gt $2 ]; then
+			range_end=$(($2-1))
+		fi
+		range_length=$((${range_end}-${range_start}+1))
+
+		# here, the filename and path matters
+		echo "upload chunk: ${chunk_index}, length: ${range_length}, bytes: ${range_start}-${range_end} "
+
+		fragment_name=$(echo ${file_name} | sed "s/\./_$((chunk_index+1))\./g")
+		fragment_path=${video_root_folder}/${file_parent}/${fragment_name}
+
+		# using command dd to split mp4 file to fragments without encoding (the fragment cannot be played)				
+		dd if="$1" of="./data/${fragment_name}" count=1 skip=${chunk_index} bs=${chunk_size} 2> /dev/null
+		query="curl -s -k -L -X PUT '${DRIVE_BASE_URI}/items/root:/${fragment_path}:/content'"	
+		query="${query} --upload-file ./data/${fragment_name}"
+		send_api_request "\tupload_file_fragment" ${fragment_name}
+
+		chunk_index=$((${chunk_index}+1))		
+		sleep 1
+	done 
+
+	color_print "GREEN" "Success: upload_large_file_by_fragments $1"	
 }
